# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Lorem Ipsum for Documents - Generate placeholder documents for testing and development workflows",
        "title": "Pseudoc API",
        "termsOfService": "https://github.com/engineervix/pseudoc",
        "contact": {
            "name": "Pseudoc Support",
            "url": "https://github.com/engineervix/pseudoc/issues"
        },
        "license": {
            "name": "BSD-3-Clause",
            "url": "https://github.com/engineervix/pseudoc/blob/main/LICENSE.md"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/generate": {
            "post": {
                "description": "Generate a document using JSON request body. Provides more flexibility than GET endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/pdf",
                    "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Generate document via POST request",
                "parameters": [
                    {
                        "description": "Document generation parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_server_handlers.GenerateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated document file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or parameters",
                        "schema": {
                            "$ref": "#/definitions/internal_server_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_server_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/generate/{type}": {
            "get": {
                "description": "Generate a document using URL path parameter and query parameters. Supports PDF, DOCX, XLSX, and random document types.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/pdf",
                    "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Generate document via GET request",
                "parameters": [
                    {
                        "enum": [
                            "pdf",
                            "docx",
                            "xlsx",
                            "random"
                        ],
                        "type": "string",
                        "example": "pdf",
                        "description": "Document type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 3,
                        "example": 5,
                        "description": "Number of pages (PDF/DOCX only)",
                        "name": "pages",
                        "in": "query"
                    },
                    {
                        "maximum": 10,
                        "minimum": 1,
                        "type": "integer",
                        "default": 2,
                        "example": 3,
                        "description": "Number of sheets (XLSX only)",
                        "name": "sheets",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "example": 42,
                        "description": "Random seed for reproducible generation",
                        "name": "seed",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "my-document",
                        "description": "Custom filename (without extension)",
                        "name": "filename",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated document file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/internal_server_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_server_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_server_handlers.ErrorResponse": {
            "description": "Standard error response format",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "Invalid document type"
                },
                "message": {
                    "type": "string",
                    "example": "Supported types: pdf, docx, xlsx, random"
                },
                "request_id": {
                    "type": "string",
                    "example": "req_123456789"
                },
                "validation_errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_server_handlers.ValidationError"
                    }
                }
            }
        },
        "internal_server_handlers.GenerateRequest": {
            "description": "Request body for document generation via POST",
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "filename": {
                    "type": "string",
                    "example": "my-document"
                },
                "pages": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1,
                    "example": 5
                },
                "seed": {
                    "type": "integer",
                    "example": 42
                },
                "sheets": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 1,
                    "example": 3
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "pdf",
                        "docx",
                        "xlsx",
                        "random"
                    ],
                    "example": "pdf"
                }
            }
        },
...
        "internal_server_handlers.ValidationError": {
            "description": "Detailed validation error information",
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "pages"
                },
                "message": {
                    "type": "string",
                    "example": "pages must be between 1 and 100"
                },
                "value": {
                    "type": "string",
                    "example": "150"
                }
            }
        },
...
    },
...
}